Reorganise gameplay settings into more sections
