+++ osu.Game/Overlays/Settings/Sections/Gameplay/AudioSettings.cs
@@ -1,0 +1,34 @@
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using osu.Framework.Allocation;
+using osu.Framework.Graphics;
+using osu.Framework.Localisation;
+using osu.Game.Configuration;
+using osu.Game.Localisation;
+
+namespace osu.Game.Overlays.Settings.Sections.Gameplay
+{
+    public class AudioSettings : SettingsSubsection
+    {
+        protected override LocalisableString Header => GameplaySettingsStrings.AudioHeader;
+
+        [BackgroundDependencyLoader]
+        private void load(OsuConfigManager config)
+        {
+            Children = new Drawable[]
+            {
+                new SettingsCheckbox
+                {
+                    LabelText = GameplaySettingsStrings.PositionalHitsounds,
+                    Current = config.GetBindable<bool>(OsuSetting.PositionalHitSounds)
+                },
+                new SettingsCheckbox
+                {
+                    LabelText = GameplaySettingsStrings.AlwaysPlayFirstComboBreak,
+                    Current = config.GetBindable<bool>(OsuSetting.AlwaysPlayFirstComboBreak)
+                },
+            };
+        }
+    }
+}
