+++ osu.Game/Overlays/Settings/Sections/Gameplay/HUDSettings.cs
@@ -1,0 +1,52 @@
+﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using osu.Framework.Allocation;
+using osu.Framework.Graphics;
+using osu.Framework.Localisation;
+using osu.Game.Configuration;
+using osu.Game.Localisation;
+using osu.Game.Rulesets.Scoring;
+
+namespace osu.Game.Overlays.Settings.Sections.Gameplay
+{
+    public class HUDSettings : SettingsSubsection
+    {
+        protected override LocalisableString Header => GameplaySettingsStrings.HUDHeader;
+
+        [BackgroundDependencyLoader]
+        private void load(OsuConfigManager config)
+        {
+            Children = new Drawable[]
+            {
+                new SettingsEnumDropdown<HUDVisibilityMode>
+                {
+                    LabelText = GameplaySettingsStrings.HUDVisibilityMode,
+                    Current = config.GetBindable<HUDVisibilityMode>(OsuSetting.HUDVisibilityMode)
+                },
+                new SettingsEnumDropdown<ScoringMode>
+                {
+                    LabelText = GameplaySettingsStrings.ScoreDisplayMode,
+                    Current = config.GetBindable<ScoringMode>(OsuSetting.ScoreDisplayMode),
+                    Keywords = new[] { "scoring" }
+                },
+                new SettingsCheckbox
+                {
+                    LabelText = GameplaySettingsStrings.ShowDifficultyGraph,
+                    Current = config.GetBindable<bool>(OsuSetting.ShowProgressGraph)
+                },
+                new SettingsCheckbox
+                {
+                    LabelText = GameplaySettingsStrings.ShowHealthDisplayWhenCantFail,
+                    Current = config.GetBindable<bool>(OsuSetting.ShowHealthDisplayWhenCantFail),
+                    Keywords = new[] { "hp", "bar" }
+                },
+                new SettingsCheckbox
+                {
+                    LabelText = GameplaySettingsStrings.AlwaysShowKeyOverlay,
+                    Current = config.GetBindable<bool>(OsuSetting.KeyOverlay)
+                },
+            };
+        }
+    }
+}
